add: nowplayingnotificationlevel
