+++ src/Component/backupper/replit.ts
@@ -20,11 +20,9 @@ import type { YmxFormat } from "../../Structure";
 import type { DataType, MusicBotBase } from "../../botBase";
 import type { JSONStatuses } from "../../types/GuildStatuses";
 
-import candyget from "candyget";
-import PQueue from "p-queue";
-
 import { IntervalBackupper } from ".";
 import { timeLoggedMethod } from "../../logger";
+import { ReplitClient } from "../ReplitDatabaseClient";
 
 export class ReplitBackupper extends IntervalBackupper {
   protected readonly db: ReplitClient = null!;
@@ -138,67 +136,5 @@ export class ReplitBackupper extends IntervalBackupper {
   }
 
   override destroy(){
-  }
-}
-
-class ReplitClient {
-  protected baseUrl: string;
-  protected queue: PQueue;
-
-  constructor(baseUrl: string){
-    this.baseUrl = baseUrl;
-
-    if(!this.baseUrl){
-      throw new Error("No URL found");
-    }
-
-    this.queue = new PQueue({
-      concurrency: 3,
-      timeout: 10e3,
-      throwOnTimeout: true,
-      intervalCap: 4,
-      interval: 10,
-    });
-  }
-
-  get(key: string, options: { raw: true }): Promise<string>;
-  get<T = any>(key: string, options?: { raw: false }): Promise<T>;
-  get(key: string, options?: { raw: boolean }){
-    return this.queue.add(async () => {
-      const shouldRaw = options?.raw || false;
-      const { body } = await candyget(`${this.baseUrl}/${key}`, "string");
-      if(!body){
-        return null;
-      }else if(shouldRaw){
-        return body;
-      }else{
-        return JSON.parse(body);
-      }
-    });
-  }
-
-  set(key: string, value: any){
-    return this.queue.add(async () => {
-      const textData = JSON.stringify(value);
-
-      const { statusCode } = await candyget.post(this.baseUrl, "empty", {
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      }, `${encodeURIComponent(key)}=${encodeURIComponent(textData)}`);
-
-      if(statusCode >= 200 && statusCode <= 299){
-        return this;
-      }else{
-        throw new Error(`Status code: ${statusCode}`);
-      }
-    });
-  }
-
-  delete(key: string){
-    return this.queue.add(async () => {
-      await candyget.delete(`${this.baseUrl}/${key}`, "empty");
-      return this;
-    });
   }
 }
